requirements.txt
discord.py>=1.7.3
python-decouple>=3.6
pymongo>=4.1.1
dnspython>=2.2.1
urllib3>=1.26.9
requests>=2.27.1